requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
PyYAML==6.0
scikit_learn==1.1.1
seaborn==0.11.1
openpyxl==3.0.10
missingno==0.5.0


